requirements.txt
python-dotenv
pipecat-ai[azure,silero,websocket]
fastapi
uvicorn